+++ NEWS.md
@@ -1,3 +1,5 @@
+# powRICLPM (development version)
+
 # powRICLPM 0.2.0
 
 ## Breaking changes